server/requirement.txt
flask
flask-cors
flask-restful
dotenv
pymongo
adafruit-io
pyjwt